迁移ipsw路由
