+++ lib/router.js
@@ -318,7 +318,7 @@ router.get('/xmu/aero/:type', lazyloadRouteHandler('./routes/universities/xmu/ae
 router.get('/ifanr/:channel?', lazyloadRouteHandler('./routes/ifanr/index'));
 
 // IPSW.me
-router.get('/ipsw/index/:ptype/:pname', lazyloadRouteHandler('./routes/ipsw/index'));
+// router.get('/ipsw/index/:ptype/:pname', lazyloadRouteHandler('./routes/ipsw/index'));
 
 // 异次元软件世界
 router.get('/iplay/home', lazyloadRouteHandler('./routes/iplay/home'));
