+++ lib/routes/ipsw/namespace.ts
@@ -1,0 +1,6 @@
+import type { Namespace } from '@/types';
+
+export const namespace: Namespace = {
+    name: 'IPSW.me',
+    url: 'ipsw.me',
+};
